Started working on Vector implementation and exceptions

+++ Matrix.h
@@ -5,5 +5,5 @@ class Matrix {
     public:
 
     private:
-        std::vector<std::vector<int>> matrix;
+        std::vector<std::vector<double>> matrix;
 };
